#1551 combine 4 files in  landing-metrics folder
