--- packages/common/src/api-arbimon/metrics/jobs-count.ts
@@ -1,7 +1,0 @@
-import { type AxiosInstance } from 'axios'
-
-export const detectJobsCount = (): string => '/legacy-api/jobs-count'
-
-export const apiArbimonGetJobsCount = async (apiClient: AxiosInstance): Promise<number | undefined> => {
-  return await apiClient.get<number>(detectJobsCount()).then(res => res.data)
-}
