--- packages/common/src/api-arbimon/metrics/recordings-species-count.ts
@@ -1,7 +1,0 @@
-import { type AxiosInstance } from 'axios'
-
-export const detectRecordingsSpeciesCount = (): string => '/legacy-api/recordings-species-count'
-
-export const apiArbimonGetRecordingsSpeciesCount = async (apiClient: AxiosInstance): Promise<number | undefined> => {
-  return await apiClient.get<number>(detectRecordingsSpeciesCount()).then(res => res.data)
-}
